Only run legacy-native-module in Release (#51494)
Summary:

A recent diff introduced the legacy-native-module.yml E2E test for maestro.

The test is consistently failing in CI in Debug mode, because it requires to navigate to the APIs tab. However, the navigation to that tab is not working in debug mode because there is a logbox on top of the tabbar that prevents the tests to tap on the button.

The test works fine in Release mode because we don't have logboxes there.

As a temporary solution to unblock CI while we fix the warnings that prevent the navigation, we are disabling the test for running in debug mode.

This will not reduce the coverage as:
- the test still runs in Release mode
- the debug tests have been added to verify the connection with metro, which is checked by the other tests in the test suite.

## Changelog:
[Internal] -

Differential Revision: D75139882