Correctly propagate RN_SERIALIZABLE_STATE to 3rd party CMake targets
Summary:
RNTester Android is currently instacrashing in OSS due to 3rd-party packages not having the `RN_SERIALIZABLE_STATE` definition.
That's because the `INTERFACE` definition is not properly propagated on the prefab boundaries.

This was happening for `react-native-popup-menu-android` and `react-native-test-library` and will also happen for Codegen libraries.

This fixes it. 3p developers with custom CMake files will also have to use the `target_compile_reactnative_options` functions to make sure the compilation flag are properly populated.

Changelog:
[Android] [Changed] - Correctly propagate RN_SERIALIZABLE_STATE to 3rd party CMake targets

Differential Revision: D75441245