Let accessible be std::optional in native
Summary:
We want to change Text so that if it has links, and no `accessible` value set, it becomes accessible by default for keyboards. To do that we need native to know if "no value is set". This means changing from a `bool` to a `std::optional<bool>`.

This diff does that and fixes any build errors. Runtime should be the same, since the default at all the callsites is false (the default before), and JS overrides will still apply.

Changelog: [Internal]

Reviewed By: javache

Differential Revision: D74914316