Avoid scratch TextPaint for Facsimile Layouts (#51827)
Summary:

TextLayoutManager has an optimization today, where it reuses a scratch TextPaint throughout layouts.

This is not safe in Facsimile, since we may mutate the paint on a Layout, after it is created, but before it is drawn.

This change makes it so that we only ever use a scratch paint, when the layout it is set on does not escape TextLayoutManager.

Changelog: [Internal]

Differential Revision: D75987605