Refactor valueFromShadowNode as createAndPromoteRuntimeShadowNodeReference (#51677)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/51677

Changelog: [internal]

This refactors `updateRuntimeShadowNodeReference` to also use the bridging function, which is now adapted to support propagating ShadowNode runtime references.

The reason for having this in the bridging method is to ensure encapsulation for the JS value, and making sure shadow node values from JS can always be converted to ShadowNode values in C++ in a predictable way.

Reviewed By: sammy-SC

Differential Revision: D75598199